Restructure codebase, add screens, add about and settings
