+++ src/SpeechJokey.kv
@@ -1,6 +1,6 @@
 #:kivy 2.3.0
 
-<MainScreen>:
+<AppScreen>:
     md_bg_color: self.theme_cls.secondaryContainerColor
 
     MDBoxLayout:
@@ -26,7 +26,7 @@
                     on_release: root.on_save_file()
 
             MDTopAppBarTitle:
-                text: app.title
+                text: root.title
                 pos_hint: {"center_x": .5, "center_y": .5}
 
             MDTopAppBarTrailingButtonContainer:
