+++ src/screens/main_screen.kv
@@ -1,0 +1,61 @@
+#:kivy 2.3.0
+
+<MainScreen>:
+    md_bg_color: self.theme_cls.secondaryContainerColor
+
+    MDBoxLayout:
+        orientation: 'vertical'
+
+        MDTopAppBar:
+            type: "small"
+            pos_hint: {"center_x": .5, "top": 1}
+            MDTopAppBarLeadingButtonContainer:
+                spacing: "4dp"
+                padding: "4dp"
+                MDActionTopAppBarButton:
+                    id: btn_menu
+                    icon: "menu"
+                    on_release: root.on_menu_open()
+                MDFabButton:
+                    id: btn_load
+                    icon: "file-upload"
+                    on_release: root.on_load_file()
+                MDFabButton:
+                    id: btn_save
+                    icon: "content-save"
+                    on_release: root.on_save_file()
+
+            MDTopAppBarTitle:
+                text: root.title
+                pos_hint: {"center_x": .5, "center_y": .5}
+
+            MDTopAppBarTrailingButtonContainer:
+                spacing: "4dp"
+                padding: "4dp"
+                MDFabButton:
+                    id: btn_play
+                    icon: "play-circle"
+                    on_press: root.on_play()
+                MDFabButton:
+                    id: btn_synthesize
+                    icon: "account-voice"
+                    on_press: root.on_synthesize()
+
+        MDScrollView:
+            adaptive_height: True
+            MDTextField:
+                id: text_main
+                mode: "filled"
+                multiline: True
+                size_hint_y: None
+                adaptive_height: True
+
+        MDBoxLayout:
+            adaptive_size: True
+            orientation: 'horizontal'
+            MDIcon:
+                icon: "information-outline"
+            MDLabel:
+                adaptive_width: True
+                id: label_status
+                text: "Uhn tiss Uhn tiss Uhn tiss"
