+++ src/settings/appsettings.kv
@@ -1,0 +1,22 @@
+#:kivy 2.3.0
+
+<AppSettings>:
+    MDBoxLayout:
+        orientation: "vertical"
+        MDTopAppBar:
+            type: "small"
+            pos_hint: {"center_x": .5, "top": 1}
+            MDTopAppBarLeadingButtonContainer:
+                spacing: "4dp"
+                padding: "4dp"
+                MDActionTopAppBarButton:
+                    id: btn_back
+                    icon: "arrow-left"
+                    on_release:
+                        root.manager.transition.direction = 'right'
+                        root.manager.current = "main"
+            MDTopAppBarTitle:
+                text: "Settings"
+                pos_hint: {"center_x": .5, "center_y": .5}
+        MDLabel:
+            text: "A setting"
